Move solve Ops into a sub-module
