+++ pytensor/tensor/_linalg/solve/rewriting.py
@@ -8,10 +8,12 @@ from pytensor.scan.op import Scan
 from pytensor.scan.rewriting import scan_seqopt1
 from pytensor.tensor._linalg.decomposition.cholesky import cholesky
 from pytensor.tensor._linalg.decomposition.lu import lu_factor
+from pytensor.tensor._linalg.solve.general import Solve, lu_solve
 from pytensor.tensor._linalg.solve.linear_control import (
     SolveBilinearDiscreteLyapunov,
     solve_discrete_lyapunov,
 )
+from pytensor.tensor._linalg.solve.psd import cho_solve
 from pytensor.tensor._linalg.solve.tridiagonal import (
     tridiagonal_lu_factor,
     tridiagonal_lu_solve,
@@ -21,7 +23,6 @@ from pytensor.tensor.blockwise import Blockwise
 from pytensor.tensor.elemwise import DimShuffle
 from pytensor.tensor.rewriting.basic import register_specialize
 from pytensor.tensor.rewriting.blockwise import blockwise_of
-from pytensor.tensor.slinalg import Solve, cho_solve, lu_solve
 from pytensor.tensor.variable import TensorVariable
